A lot of problems kinda fixed.
I faced numerous bugs, errors while trying to create a jar and an exe file for my program. I got a lot of help from internet. This is the one of the last commits i will be doing for this version. The code is quite messy and its probably not optimal. But thats future me's problem. Right now I am happy with the program. Will publish the program later.